requirements.txt
numpy==1.16.4
Pillow==6.1.0
matplotlib==3.0.3
opencv-python==4.4.0.46
